Moved soundplayer to resources package
