+++ app/src/main/java/knightec/bicyclepriority/presentation/MainActivity.kt
@@ -11,11 +11,9 @@ import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.location.LocationManager
-import android.os.Build
 import android.os.Bundle
 import android.os.VibrationEffect
 import android.os.Vibrator
-import android.os.VibratorManager
 import android.provider.Settings
 import android.widget.Toast
 import androidx.activity.ComponentActivity
@@ -23,18 +21,13 @@ import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
 import androidx.wear.compose.material.*
-import androidx.wear.compose.material.*
-import knightec.bicyclepriority.presentation.repository.SoundPlayer
+import knightec.bicyclepriority.presentation.utilities.SoundPlayer
 import knightec.bicyclepriority.presentation.theme.BicyclePriorityTheme
 import knightec.bicyclepriority.presentation.view.LocationView
 import knightec.bicyclepriority.presentation.view.TrafficLightView
@@ -56,8 +49,6 @@ class MainActivity : ComponentActivity() {
         trafficLightView = TrafficLightView(trafficLightViewModel)
         prepLocationUpdates()
         createLocationView()
-
-
 
         val timings: LongArray = longArrayOf(50, 100, 50, 100, 50)
         val amplitudes: IntArray = intArrayOf(0, 255, 0, 255, 0)
